third_party/WebKit/LayoutTests/http/tests/devtools/console/console-retain-autocomplete-on-typing-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Verify that console does not hide autocomplete during typing.


Running: testSummonSuggestBox
Suggestions shown.

Running: testTypeText
SUCCESS: suggestbox is not hidden during typing.
